Split Zapper provider modules
